Fixed bug where if teh player had more ammo in the clip then the total ammo the player would be unable to pick up more ammo

+++ Assets/Scripts/Boss.cs
@@ -78,11 +78,11 @@ public class Boss : EnemyAI
             }
             else
             {
-                base.TakeDamage((damage / 2));
+                base.TakeDamage((int)((float)damage / 2f));
                 
             }
+
             canShootGrenade = true; 
-            
         }
         else
         {
